Don't concatenate levels to variables that only have one in CAs
Variables that only have one level are most likely boolean, even
if their are not identified as such (bad importation by users).
Thus, there's no point in concatenating the name of their only
level, which sounds weird in plots.

This is mostly useful for users that have manually created dummy
variables, but with only TRUE/NA values: only the average point for
TRUE is interesting, so FALSE isn't used, and NA is never shown.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr-tms@10 edb9625f-4e0d-4859-8d74-9fd3b1da38cb