Add support for more file types
On Windows, RODBC allows us to read Excel and Access files. On
Linux, ROpenOffice can read ODS files. The former isn't always
available on Mac OS (even if a driver can be installed) and almost
never on Linux, so disable it there. The latter isn't available
as binaries ATM, so we only ask to install it on Linux. This
means Mac OS cannot read these file types.

When a file type is not supported, show it nonetheless in the
types list, but give an error message when trying to load a file:
this gives more information to the user about why this type is not
supported. Else, supported types would vary without explanation
between platforms.

By the way, also accept tab-separated values files as .tsv, and
dBase files as .dbf, since it's pretty easy. These formats don't
specify character encodings, so they are not robust if moved
across computers.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr-tms@34 edb9625f-4e0d-4859-8d74-9fd3b1da38cb