Add word length measures
New submenu to analyse word lengths, based on the existing terms
vocabulary menu and dialogs. Compared to it, the improvements are that
it uses lattice for barcharts, to ensure legends are always correctly
drawn. Vocabulary will be changed too.

We now clean objects when loading a new corpus: this allows the code
to assume that already existing document-term matrices can be safely
used instead of computing a new one, which is very useful for the
word lengths menu.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr-tms@69 edb9625f-4e0d-4859-8d74-9fd3b1da38cb