Factiva: Adapt to upcoming changes in tm and fix bugs with paragraphs
tm 0.5-7.2 uses XML internal nodes, which fixes a few problems with 
the parsed XML, but we need to adapt a little to this by using xmlChildren().
Also fix a bug with concatenating tailParagraphs: add a new line to avoid
words from the end and the beginning of the next paragraph from being
concatenated.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr-tms@78 edb9625f-4e0d-4859-8d74-9fd3b1da38cb