Create empty variable in corpusVars as factor and remove levels in clustering
If we use NA without details, the column is set up as logical(),
and further attemps to fill it with anything different from TRUE/FALSE
will fail without any warning. This is made even more confusing by
the fact that (at least on Linux) logical variables appear as Character
in the editor.

This changes the behavior introduced in ea7bb4b (r143): using "" was not
optimal either since it means an empty factor level is created, which is
shown everywhere after that.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr-tms@179 edb9625f-4e0d-4859-8d74-9fd3b1da38cb