Add HTML output framework
We need to adapt tables so that we have the needed information:
what table was produced last, and what title should it use. Also
improve default titles, and make them separate from plotting options.

Remove the uses of HTML2clip, which did not work except in rare some
cases (text is not marked as HTML, and is pasted as plain text in most
word processors).

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rcmdr-tms@187 edb9625f-4e0d-4859-8d74-9fd3b1da38cb